Refactor the code in smaller modules.
